Fixed the nist-exit cgi call, and updated the working group list.
The bad cgi line breaks the link to nanohub from the links on the left
hand edge of the manual pages.  This was reported by the nanohub
people.

I haven't actually rebuilt the manual pages with this new menu.html
file, because to do that I need to be able to run the correct old
version of oof2 using python2.7, and for some reason that's not
working.  If the problem persists after the manual is updated for
python3, I'll fix it then.