Formatting changes.
fonts.xsl now uses <code> tags for Registration and RegisteredClass
names, instead of <tt>.  For some reason <code> is rendered a lot
more nicely, at least on Chrome.  The change makes the names more
consistent with other code-like objects.

Fixed a bug in oof.css, changed the background color for figure
captions (it was too close to a color I was using in a figure),
and did a TODO about font sizes for <code> objects.