Starting to fix all of the displayed images in the manual.
They had been exported at 72 dpi from OmniGraffle, which made them
show up the correct size when the html was displayed, but they were
blurry.  It turns out that hte screenshots I was feeding into
OmniGraffle were 144 dpi.  Exporting at 144 dpi from OmniGraffle
made the images twice as big in html.  Adding scale="50" to
the docbook imagedata tags should have fixed that, but didn't.
The correct scale factor was hard to figure out and differed
for each image. (The displayed size was sometimes a nonlinear function of
the scale factor.)  What works is to use contentwidth="100px", where
the size is the HALF the width in pixels of the OmniGraffle output.  So
not only do I have to re-export the images at 144 dpi, but I have
to edit all of the imagedata tags.

I think this is a bug in some part of the xml processing machinery
that I'm using, but it's all a decade old, and the process of figuring
it all out again so that I can use a new version is surely going to
take more time than regenerating the images.  I tried just dropping
in a newer version of saxon.jar, and it didn't work.