Going through all of the discussion files, checking for things that need fixing, especially images.
I'm searching for and deleting the old CVS headers as I go, as it's
a pretty good way of finding the files that I haven't looked at in
a while.  Also I'm checking that the email address in the headers
is correct. That's why so many files have changed.  There are more
to come.