Working on code overview and starting on class specifics.
All classes will have an entity defined in man_oof2.xml
so that "&classname;" will be replaced by a link to the
reference page for the class. The names are case-sensitive,
so that "&node;" still links to section on nodes in
CH_concepts.xml, but "&Node;" links to the reference page
for the node class.