Reorganized the reference pages, using the <reference> tag and <part>.
Also started working on Property documentation, and continued
to tweak indices and iterators.  That should have been in a separate
commit from the reorganization, but I can't easily separate it now.

I still don't understand how sgml-parent-document works.  I can't
find a setting that says the xml is valid and also doesn't spew
a warning when emacs loads a file.  The values in the CODEREF
files are now inconsistent, but it doesn't seem to matter.