Changes to use MathML instead of latex and dvi2bitmap.
The Makefile behaves just like it did before, processing the xml
files and producing html files along with tex-math-equations.tex
and tex-math-inlines.tex.  But instead of running latex and dvi2bitmap
to create gif files for the equations, it runs converteqns.py, which
runs latexmlc to convert latex to mathml and replaces the <img>
tags in the html files with the generated mathml code.

This version only works if we use the docbook xhtml templates.
Using xhtml causes all sorts of other problems, so I will
experiment more to see if I can get it work with html instead.
The switch to xhtml was done in xsl/oofchunk.xsl and xsl/oofhtml.xsl
by changing which docbook files are imported/included.

The reason for using xhtml instead of html is that the <img> tags
aren't properly closed in html, and that makes it hard for
converteqns.py to find and replace them.