Changed the manual's Makefile to make it easier to build different manuals from different versions of oof2.
I was trying to compare the manuals generated by oof2 in the
fix-menu-options branch to the manual created by the develop branch
at the point where fix-menu-options diverged.  fix-menu-options was
needed because some menu items weren't appearing in the manual.
Hacking in the changes required to create a separate manual was
error prone.  Now it's simpler.

Instead of relying on a mix of Make variables, symbolic links, and
environment variables, the Makefile's behavior is now determined
completely by variables that are set within the script.